+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.6.4]
 - Fix convert water
 - Improve particles convert and merge Java 1.13 and 1.14 variants
+- Fix parse `pack.mcmeta` files with 'UFT8-BOOM'
 
 ## [1.6.3]
 - Fix `publish` script