+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [1.6.5]
+- Fix painting are warped
+
 ## [1.6.4]
 - Fix convert water
 - Improve particles convert and merge Java 1.13 and 1.14 variants